Swagger

+++ src/main/java/com/example/demo/dao/OrderDao.java
@@ -14,6 +14,6 @@ public interface OrderDao extends JpaRepository<OrderEntity, Long> {
     @Transactional
     @Modifying
     @Query(value="UPDATE orderItem SET isCancel = True WHERE order_id=?1")
-    void updateOrder(int orderId);
+    void updateOrder(Long orderId);
 
 }
